+++ jquery.reel.js
@@ -24,7 +24,7 @@
  * jQuery Reel
  * http://reel360.org
  * Version: 1.3.1-devel
- * Updated: 2014-02-28
+ * Updated: 2014-03-24
  *
  * Requires jQuery 1.6.2 or higher
  */
@@ -1134,6 +1134,7 @@
                   set(_ticks_, -1);
                   set(_annotations_, opt.annotations || $overlay.unbind(dot(_annotations_)) && {});
                   set(_ratio_, 1);
+                  set(_monitor_, '');
                   set(_backup_, {
                     attr: attr,
                     data: data
@@ -1239,8 +1240,7 @@
                     opt.hint && $area.attr('title', opt.hint);
                     opt.indicator && $overlay.append(indicator('x'));
                     multirow && opt.indicator && $overlay.append(indicator('y'));
-                    opt.monitor && $overlay.append($monitor= $(tag(_div_), { 'class': monitor_klass }))
-                                && css(___+dot(monitor_klass), { position: _absolute_, left: 0, top: 0 });
+                    opt.monitor && $overlay.append(monitor())
                   },
 
                   // ### `preload` Event ######
@@ -1821,6 +1821,17 @@
                   // more CPU resources than the entire Reel scene. Use them for development only.
                   //
 
+                  // -----------------
+                  // Debugging Monitor
+                  // -----------------
+                  //
+                  // With [`monitor`](#monitor-Option) option set, a simple text node in the upper left corner
+                  // of the scene is continuously updated with the actual value of given monitored data field.
+                  // 
+                  monitorChange: function(e, nil, value){
+                    monitor.$.text(value);
+                  },
+
                   // -----------
                   // Annotations
                   // -----------
@@ -2051,7 +2062,7 @@
                     if (braking) var
                       braked= velocity - (get(_brake_) / leader_tempo * braking),
                       velocity= set(_velocity_, braked > 0.1 ? braked : (braking= operated= 0))
-                    monitor && $monitor.text(get(monitor));
+                    monitor && set(_monitor_, get(monitor)+__);
                     velocity && braking++;
                     operated && operated++;
                     to_bias(0);
@@ -2119,7 +2130,14 @@
 
               // - Constructors of UI elements
               //
-              $monitor= $(),
+              monitor= function(){
+                css(___+dot(monitor_klass), {
+                  position: _absolute_,
+                  left: 0,
+                  top: 0
+                });
+                return monitor.$= $(tag(_div_), { 'class': monitor_klass })
+              },
               preloader= function(){
                 css(___+dot(preloader_klass), {
                   position: _absolute_,
@@ -2717,7 +2735,7 @@
     _height_= 'height', _hi_= 'hi', _hidden_= 'hidden',
     _image_= 'image', _images_= 'images',
     _lo_= 'lo', _loading_= 'loading',
-    _mouse_= 'mouse',
+    _monitor_= 'monitor', _mouse_= 'mouse',
     _opening_= 'opening', _opening_ticks_= _opening_+'_ticks', _options_= 'options',
     _playing_= 'playing', _preloaded_= 'preloaded',
     _ratio_= 'ratio', _reeling_= 'reeling', _reeled_= 'reeled', _responsive_= 'responsive', _revolution_= 'revolution',