+++ CHANGELOG.markdown
@@ -9,6 +9,7 @@ Development
 * Negative speed now rolling in correct direction (GH-241).
 * Compatibility with Visual Studio Windows 8 apps added (GH-263).
 * Cache teardown removal fixed (GH-265).
+* Improved performance of debugging monitor.
 
 
 Version 1.3